+++ res/views/components/button.blade.php
@@ -1,10 +1,18 @@
 @isset($href)
     @tag('a', $attributes->merge(compact('href')))
-        <s-text :value="$text">{{ $slot }}</s-text>
+        @isset($text)
+            <s-text :value="$text"/>
+        @else
+            {{ $slot }}
+        @endisset
     @endtag
 @elseif(isset($text) || $slot->isNotEmpty())
     @tag('button', $attributes->merge(compact('type')))
-        <s-text :value="$text">{{ $slot }}</s-text>
+        @isset($text)
+            <s-text :value="$text"/>
+        @else
+            {{ $slot }}
+        @endisset
     @endtag
 @else
     @tag('input', $attributes->merge(compact('type', 'name', 'value')))